Split Hive attach payload helper
